// src/types/WeatherAlert.ts
export interface WeatherAlert {
  id: string;
  type: string;
  areaDescription: string;
  category: string;
  certainty: string;
  description: string;
  effective: string;
  expires: string;
  event: string;
  headline: string;
  instruction: string;
  messageType: string;
  response: string;
  senderName: string;
  severity: string;
  statement: string;
  status: string;
  urgency: string;
}